check the uid before using it.
2003-07-24  Rodrigo Moya <rodrigo@ximian.com>

	* gui/e-cal-view.c (delete_event): check the uid before using it.

svn path=/trunk/; revision=21944

+++ calendar/gui/e-cal-view.c
@@ -593,6 +593,8 @@ delete_event (ECalView *cal_view, CalComponent *comp)
 					cal_view->priv->client, NULL);
 
 		cal_component_get_uid (comp, &uid);
+		if (!uid || !*uid)
+			return;
 
 		delete_error_dialog (
 			cal_client_remove_object (cal_view->priv->client, uid), CAL_COMPONENT_EVENT);
